q_and_a/requirements.txt
fastapi
uvicorn[standard]
dedalus-labs
python-dotenv
pydantic
fastmcp
openai

